# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Mercury API
  version: 0.0.1
  description: REST API contract for Mercury messaging services.
servers:
  - url: /
paths:
  /api/v1/auth/token:
    post:
      tags:
        - auth
      summary: Generate a session token (M2M-only, no uid)
      description: >-
        Implemented by AuthTokenController. Checks the request body's alias/password
        against Mercury's registered login-client registry (umdc.mercury.login-clients)
        and mints a Mercury-signed session-token. The resulting token carries no uid
        claim — it only authenticates the calling backend service, not an end user — so
        it's only useful against pure M2M endpoints (/api/v1/mail,
        /api/v1/verification-code). User-scoped endpoints (Template Management,
        Campaign's getByApplication/updateCampaign/toggleCampaign/deleteCampaign) do NOT
        use this token: the caller sends the end user's backbone-rest session token
        directly in the session-token header.
      operationId: accessToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Successfully generated session token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Blank alias.
        '401':
          description: Alias/password not in the registered login-client registry.
        '429':
          description: Too many failed attempts for this alias and address; see the Retry-After header.
        '406':
          description: Token generation unexpectedly produced a blank token.
...
  /api/v1/campaigns:
    post:
      tags:
        - campaigns
      summary: Create a new campaign
      description: Creates a new messaging campaign and publishes messages to recipients via the specified channel.
      operationId: createCampaign
      parameters:
        - name: session-token
          in: header
          required: true
          description: The end user's backbone-rest session token.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Campaign created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCampaignResponse'
        '400':
          description: Invalid request payload.
        '401':
          description: Missing, expired, tampered or revoked session-token.
        '403':
          description: The request userId does not match the session-token user.
        '422':
          description: Channel type not found or disabled.
    get:
      tags:
        - campaigns
      summary: Get campaigns by application for current user
      description: Retrieves campaigns filtered by application id for the authenticated user. User id is obtained from the session token header.
      operationId: getCampaignsByApplication
      parameters:
        - name: applicationId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: session-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Campaigns retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignDetailResponse'
        '400':
          description: Invalid request parameters.
        '401':
          description: Invalid or missing authentication token.
        '403':
          description: Caller lacks permission to view these campaigns.
  /api/v1/campaigns/{id}:
    get:
      tags:
        - campaigns
      summary: Get campaign by ID
      description: Retrieves the full details of a campaign identified by its UUID. Only the campaign's owner (the session-token user) may read it.
      operationId: getCampaignById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: session-token
          in: header
          required: true
          description: The end user's backbone-rest session token.
          schema:
            type: string
      responses:
        '200':
          description: Campaign found and returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignDetailResponse'
        '400':
          description: Invalid UUID format supplied for id.
        '401':
          description: Invalid or missing authentication token.
        '403':
          description: Caller lacks permission to view this campaign.
        '404':
          description: Campaign with the given id does not exist.
        '500':
          description: Unexpected internal error.
    put:
      tags:
        - campaigns
      summary: Update campaign by ID
      description: Update mutable fields of a campaign identified by UUID.
      operationId: updateCampaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: session-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCampaignRequest'
      responses:
        '200':
          description: Campaign updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignDetailResponse'
        '400':
          description: Invalid request payload.
        '401':
          description: Invalid or missing authentication token.
        '403':
          description: Caller lacks permission to update this campaign.
        '404':
          description: Campaign not found.
        '422':
          description: Business rule violation.
...
  /api/v1/campaigns/{id}/toggle:
    patch:
      tags:
        - campaigns
      summary: Enable or disable a campaign
      description: Enable or disable (pause/resume) a campaign by id.
      operationId: toggleCampaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
        - name: session-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Campaign toggled successfully.
        '400':
          description: Invalid parameters.
        '401':
          description: Invalid or missing authentication token.
        '403':
          description: Caller lacks permission.
        '404':
          description: Campaign not found.
        '422':
          description: Business rule violation.
...
components:
  schemas:
    AuthRequest:
      type: object
      description: >-
        Credentials of a registered Mercury login client (umdc.mercury.login-clients)
        — not an end user's credentials.
      required:
        - alias
        - password
      properties:
        alias:
          type: string
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: The newly minted Mercury-signed session-token.
    CreateCampaignRequest:
      type: object
      required:
        - name
        - channelTypeCode
        - templateId
        - userId
        - recipients
        - status
        - applicationId
      properties:
        name:
          type: string
        channelTypeCode:
          type: string
        templateId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        recipients:
          type: array
          items:
            $ref: '#/components/schemas/RecipientTO'
        templateParams:
          type: object
          additionalProperties: true
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
        applicationId:
          type: string
          format: uuid
      example:
        name: "Welcome Campaign"
        channelTypeCode: "EMAIL"
        templateId: "00000000-0000-0000-0000-000000000001"
        userId: "00000000-0000-0000-0000-000000000002"
        recipients:
          - identifier: "user@example.com"
            name: "Jane Doe"
        templateParams:
          firstName: "Jane"
        scheduledAt: "2026-03-20T10:00:00Z"
        status: "SCHEDULED"
        applicationId: "00000000-0000-0000-0000-000000000003"
    CreateCampaignResponse:
      type: object
      properties:
        campaignId:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        totalRecipients:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        scheduledAt:
          type: string
          format: date-time
          nullable: true
    RecipientTO:
      type: object
      required:
        - identifier
      properties:
        identifier:
          type: string
          description: Recipient identifier such as email, phone, or chat ID.
        name:
          type: string
          nullable: true
        customParams:
          type: object
          additionalProperties: true
...
    CampaignDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        channelType:
          type: string
        templateId:
          type: string
          format: uuid
        status:
          type: string
        totalRecipients:
          type: integer
          format: int32
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
    UpdateCampaignRequest:
      type: object
      properties:
        name:
          type: string
        templateId:
          type: string
          format: uuid
        recipients:
          type: array
          items:
            $ref: '#/components/schemas/RecipientTO'
        templateParams:
          type: object
          additionalProperties: true
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
        applicationId:
          type: string
          format: uuid
      example:
        name: "Updated Campaign Name"
        scheduledAt: "2026-03-20T13:00:00Z"
        status: "SCHEDULED"
...